+++ tower-http/CHANGELOG.md
@@ -10,10 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix the bug that `ServeDir` can not handle percent encoded uri path correctly.
 - Enhancement: avoid using permanent redirection when redirect directory path without a trailing slash to the one has
 - Fix a [bug](https://github.com/tower-rs/tower-http/issues/121) which happens when `append_index_html_on_directories` is set to `false` in `ServeDir`.
+- Add `Cors` for setting [CORS] headers.
 
 ## Breaking changes
 
-- Add `Cors` for setting [CORS] headers.
+None.
 
 # 0.1.1 (July 2, 2021)